requirements.txt
ConfigArgParse==0.11.0
python-gitlab==0.16
py-trello==0.6.1
